django_tdd/requirements.txt
django~=3.1.3
djangorestframework~=3.12.1
psycopg2-binary~=2.8.6
pytest~=6.1.2
pytest-django~=4.1.0
